+++ go/libraries/doltcore/sqle/statsdb/noms.go
@@ -1,13 +1,25 @@
 package statsdb
 
 import (
+	"errors"
 	"fmt"
 	"github.com/dolthub/dolt/go/libraries/doltcore/dbfactory"
+	"github.com/dolthub/dolt/go/libraries/doltcore/schema"
+	"github.com/dolthub/dolt/go/libraries/doltcore/sqle/dsess"
+	"github.com/dolthub/dolt/go/libraries/doltcore/sqle/dtables"
+	"github.com/dolthub/dolt/go/libraries/doltcore/sqle/statspro"
 	"github.com/dolthub/dolt/go/libraries/utils/earl"
 	"github.com/dolthub/dolt/go/libraries/utils/filesys"
+	"github.com/dolthub/dolt/go/store/hash"
+	"github.com/dolthub/dolt/go/store/prolly"
 	"github.com/dolthub/go-mysql-server/sql"
+	"github.com/dolthub/go-mysql-server/sql/stats"
+	"io"
 	"os"
 	"path/filepath"
+	"strconv"
+	"strings"
+	"time"
 )
 
 type NomsStatsFactory struct{}
@@ -49,10 +61,156 @@ func (n NomsStats) ListStatQuals() []sql.StatQualifier {
 	panic("implement me")
 }
 
-func (n NomsStats) Load() error {
-	//TODO implement me
+func (n NomsStats) Load(branch string) error {
+	// get the ref for this branch
+	// read the map
+	// load the data
 	panic("implement me")
 }
+
+func loadStats(ctx *sql.Context, db dsess.SqlDatabase,  m prolly.Map) (map[sql.StatQualifier]*statspro.DoltStats, error) {
+	qualToStats := make(map[sql.StatQualifier]*statspro.DoltStats)
+
+	iter, err := dtables.NewStatsIter(ctx, m)
+	if err != nil {
+		return nil, err
+	}
+	currentStat := statspro.NewDoltStats()
+	var lowerBound sql.Row
+	for {
+		row, err := iter.Next(ctx)
+		if errors.Is(err, io.EOF) {
+			break
+		} else if err != nil {
+			return nil, err
+		}
+
+		position := row[schema.StatsPositionTag].(int)
+
+		// deserialize K, V
+		dbName := row[schema.StatsDbTag].(string)
+		tableName := row[schema.StatsTableTag].(string)
+		indexName := row[schema.StatsIndexTag].(string)
+		_ = row[schema.StatsVersionTag]
+		commit := hash.Parse(row[schema.StatsCommitHashTag].(string))
+		rowCount := row[schema.StatsRowCountTag].(uint64)
+		distinctCount := row[schema.StatsDistinctCountTag].(uint64)
+		nullCount := row[schema.StatsNullCountTag].(uint64)
+		columns := strings.Split(row[schema.StatsColumnsTag].(string), ",")
+		typesStr := row[schema.StatsTypesTag].(string)
+		boundRowStr := row[schema.StatsUpperBoundTag].(string)
+		upperBoundCnt := row[schema.StatsUpperBoundCntTag].(uint64)
+		createdAt := row[schema.StatsCreatedAtTag].(time.Time)
+
+		typs := strings.Split(typesStr, ",")
+		for i, t := range typs {
+			typs[i] = strings.TrimSpace(t)
+		}
+
+		numMcvs := schema.StatsMcvCountsTag - schema.StatsMcv1Tag
+
+		mcvCountsStr := strings.Split(row[schema.StatsMcvCountsTag].(string), ",")
+		mcvCnts := make([]uint64, numMcvs)
+		for i, v := range mcvCountsStr {
+			val, err := strconv.Atoi(v)
+			if err != nil {
+				return nil, err
+			}
+			mcvCnts[i] = uint64(val)
+		}
+
+		mcvs := make([]sql.Row, numMcvs)
+		for i, v := range row[schema.StatsMcv1Tag:schema.StatsMcvCountsTag] {
+			if v != nil {
+				row, err := iter.ParseRow(v.(string))
+				if err != nil {
+					return nil, err
+				}
+				mcvs[i] = row
+			}
+		}
+
+		for i, v := range mcvCnts {
+			if v == 0 {
+				mcvs = mcvs[:i]
+				mcvCnts = mcvCnts[:i]
+				break
+			}
+		}
+
+		boundRow, err := iter.ParseRow(boundRowStr)
+		if err != nil {
+			return nil, err
+		}
+
+		qual := sql.NewStatQualifier(dbName, tableName, indexName)
+		if currentStat.Qual.String() != qual.String() {
+			if !currentStat.Qual.Empty() {
+				currentStat.LowerBound, err = loadLowerBound(ctx, currentStat.Qual)
+				if err != nil {
+					return nil, err
+				}
+				fds, colSet, err := loadFuncDeps(ctx, db, currentStat.Qual)
+				if err != nil {
+					return nil, err
+				}
+				currentStat.fds = fds
+				currentStat.colSet = colSet
+				currentStat.updateActive()
+				dbStat.stats[currentStat.Qual] = currentStat
+			}
+
+			currentStat = NewDoltStats()
+			currentStat.Qual = qual
+			currentStat.Columns = columns
+			currentStat.LowerBound = lowerBound
+		}
+
+		if currentStat.Histogram == nil {
+			currentStat.Types, err = stats.ParseTypeStrings(typs)
+			if err != nil {
+				return nil, err
+			}
+			currentStat.Qual = qual
+		}
+
+		bucket := DoltBucket{
+			Chunk:         commit,
+			RowCount:      uint64(rowCount),
+			DistinctCount: uint64(distinctCount),
+			NullCount:     uint64(nullCount),
+			CreatedAt:     createdAt,
+			Mcvs:          mcvs,
+			McvCount:      mcvCnts,
+			BoundCount:    upperBoundCnt,
+			UpperBound:    boundRow,
+		}
+
+		currentStat.active[commit] = position
+		currentStat.Histogram = append(currentStat.Histogram, bucket)
+		currentStat.RowCount += uint64(rowCount)
+		currentStat.DistinctCount += uint64(distinctCount)
+		currentStat.NullCount += uint64(rowCount)
+		if currentStat.CreatedAt.Before(createdAt) {
+			currentStat.CreatedAt = createdAt
+		}
+	}
+	currentStat.LowerBound, err = loadLowerBound(ctx, currentStat.Qual)
+	if err != nil {
+		return nil, err
+	}
+	fds, colSet, err := loadFuncDeps(ctx, db, currentStat.Qual)
+	if err != nil {
+		return nil, err
+	}
+	currentStat.fds = fds
+	currentStat.colSet = colSet
+	currentStat.updateActive()
+	dbStat.setIndexStats(currentStat.Qual, currentStat)
+	dbStat.stats[currentStat.Qual] = currentStat
+	return dbStat, nil
+}
+
 
 func (n NomsStats) HasStat() {
 	//TODO implement me
